requirements.txt
argh==0.26.2
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
evernote3==1.25.13
httplib2==0.14.0
idna==2.8
oauth2==1.9.0.post1
oauthlib==3.1.0
pathtools==0.1.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.7
watchdog==0.8.2
pandas==0.24.2
numpy==1.16.4
matplotlib==3.1.0